Mount Deep Speech2 Architecture
