+++ kospeech/checkpoint/checkpoint.py
@@ -11,7 +11,7 @@ import torch
 import torch.nn as nn
 from kospeech.utils import logger
 from kospeech.data.data_loader import SpectrogramDataset
-from kospeech.models.las.las import ListenAttendSpell
+from kospeech.models.las.model import ListenAttendSpell
 from kospeech.optim.optimizer import Optimizer
