+++ kospeech/models/extractor.py
@@ -1,0 +1,74 @@
+import torch.nn as nn
+from torch import Tensor
+from typing import Any, Optional
+from kospeech.models.modules import CNNExtractor, MaskConv
+
+
+class DeepSpeech2Extractor(CNNExtractor):
+    """
+    DeepSpeech2 extractor for automatic speech recognition described in
+    "Deep Speech 2: End-to-End Speech Recognition in English and Mandarin" paper
+    - https://arxiv.org/abs/1512.02595
+    """
+    def __init__(self, activation: str = 'hardtanh', mask_conv: bool = False) -> None:
+        super(DeepSpeech2Extractor, self).__init__(activation)
+        self.mask_conv = mask_conv
+        self.conv = nn.Sequential(
+            nn.Conv2d(1, 32, kernel_size=(41, 11), stride=(2, 2), padding=(20, 5), bias=False),
+            nn.BatchNorm2d(32),
+            self.activation,
+            nn.Conv2d(32, 32, kernel_size=(21, 11), stride=(2, 1), padding=(10, 5), bias=False),
+            nn.BatchNorm2d(32),
+            self.activation
+        )
+        if mask_conv:
+            self.conv = MaskConv(self.conv)
+
+    def forward(self, inputs: Tensor, input_lengths: Tensor) -> Optional[Any]:
+        if self.mask_conv:
+            conv_feat, seq_lengths = self.conv(inputs, input_lengths)
+            output = conv_feat, seq_lengths
+        else:
+            conv_feat = self.conv(inputs)
+            output = conv_feat
+
+        return output
+
+
+class VGGExtractor(CNNExtractor):
+    """
+    VGG extractor for automatic speech recognition described in
+    "Advances in Joint CTC-Attention based End-to-End Speech Recognition with a Deep CNN Encoder and RNN-LM" paper
+    - https://arxiv.org/pdf/1706.02737.pdf
+    """
+    def __init__(self, activation: str, mask_conv: bool):
+        super(VGGExtractor, self).__init__(activation)
+        self.mask_conv = mask_conv
+        self.conv = nn.Sequential(
+            nn.Conv2d(1, 64, kernel_size=3, stride=1, padding=1, bias=False),
+            nn.BatchNorm2d(num_features=64),
+            self.activation,
+            nn.Conv2d(64, 64, kernel_size=3, stride=1, padding=1, bias=False),
+            nn.BatchNorm2d(num_features=64),
+            self.activation,
+            nn.MaxPool2d(2, stride=2),
+            nn.Conv2d(64, 128, kernel_size=3, stride=1, padding=1, bias=False),
+            nn.BatchNorm2d(num_features=128),
+            self.activation,
+            nn.Conv2d(128, 128, kernel_size=3, stride=1, padding=1, bias=False),
+            nn.BatchNorm2d(num_features=128),
+            self.activation,
+            nn.MaxPool2d(2, stride=2)
+        )
+        if mask_conv:
+            self.conv = MaskConv(self.conv)
+
+    def forward(self, inputs: Tensor, input_lengths: Tensor) -> Optional[Any]:
+        if self.mask_conv:
+            conv_feat, seq_lengths = self.conv(inputs, input_lengths)
+            output = conv_feat, seq_lengths
+        else:
+            conv_feat = self.conv(inputs)
+            output = conv_feat
+
+        return output
