+++ kospeech/optim/lr_scheduler/tri_stage_lr_scheduler.py
@@ -3,31 +3,9 @@
 # @ArXiv : KoSpeech: Open-Source Toolkit for End-to-End Korean Speech Recognition
 # This source code is licensed under the Apache 2.0 License license found in the
 # LICENSE file in the root directory of this source tree.
+
 import math
-
-
-class LearningRateScheduler(object):
-    """
-    Provides inteface of learning rate scheduler.
-
-    Note:
-        Do not use this class directly, use one of the sub classes.
-    """
-    def __init__(self, optimizer, init_lr):
-        self.optimizer = optimizer
-        self.init_lr = init_lr
-
-    def step(self, *args, **kwargs):
-        raise NotImplementedError
-
-    @staticmethod
-    def set_lr(optimizer, lr):
-        for g in optimizer.param_groups:
-            g['lr'] = lr
-
-    def get_lr(self):
-        for g in self.optimizer.param_groups:
-            return g['lr']
+from kospeech.optim.lr_scheduler.lr_scheduler import LearningRateScheduler
 
 
 class TriStageLRScheduler(LearningRateScheduler):
